Add inter-process and thread locks around any code that access the git repository.
This can cause startup to be a little slower as each server process that starts up is being made to wait on this lock.  But in practice I find it's only slow the very first time the plugin is used, and the gitolite-admin repo doesn't exist yet.  After the server is up again it adds no noticeable overhead unless two users try to update their keys simultaneously.  A possible argument to doing this outside the webserver process.  It's not a very likely scenario though so we'll see if has any real world impact or not.