Initial cleanup to the GitMerger component.
Switched it over to use GenericTableProvider for its table--this
requires a small DB upgrade (just to insert the associated schema
version into the 'system' table).

This then also gets rid of the stuff where the entire table is
created/destroyed on each cache access, and instead just deletes
individual entries as needed.  This still needs some work though; there
appear to be bugs.