Add new [sage_trac:status_badges] config so we can easily add/modify these through the config file.
The current implement meets all the needs of the existing patchbot status badges,
as well as the new one proposed in https://github.com/sagemath/sage_trac_plugin/pull/22

We could further extend the feature if more needs arise.

This also obsoletes the patchbot_url option in the [sage_trac] config, since it
was currently only used for generating the status badges.  But I kept it for now.